Mon, Mar 14, 2022, 7:55 AM +00:00

+++ filter-more/helpers.c
@@ -87,9 +87,9 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
                 }
                 else if (i == height - 1)
                 {
-                    BYTE red = 0;
-                    BYTE green = 0;
-                    BYTE blue = 0;
+                    double red = 0;
+                    double green = 0;
+                    double blue = 0;
 
                     s[0] = image[i][j + 1];
                     s[1] = image[i - 1][j + 1];
@@ -109,9 +109,9 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
                 }
                 else
                 {
-                    BYTE red = 0;
-                    BYTE green = 0;
-                    BYTE blue = 0;
+                    double red = 0;
+                    double green = 0;
+                    double blue = 0;
 
                     s[0] = image[i][j + 1];
                     s[1] = image[i + 1][j + 1];
@@ -136,9 +136,9 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
             {
                 if (i == 0)
                 {
-                    BYTE red = 0;
-                    BYTE green = 0;
-                    BYTE blue = 0;
+                    double red = 0;
+                    double green = 0;
+                    double blue = 0;
 
                     s[0] = image[i][j - 1];
                     s[1] = image[i + 1][j - 1];
@@ -158,9 +158,9 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
                 }
                 else if (i == height - 1)
                 {
-                    BYTE red = 0;
-                    BYTE green = 0;
-                    BYTE blue = 0;
+                    double red = 0;
+                    double green = 0;
+                    double blue = 0;
 
                     s[0] = image[i][j - 1];
                     s[1] = image[i - 1][j - 1];
@@ -180,9 +180,9 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
                 }
                 else
                 {
-                    BYTE red = 0;
-                    BYTE green = 0;
-                    BYTE blue = 0;
+                    double red = 0;
+                    double green = 0;
+                    double blue = 0;
 
                     s[0] = image[i][j - 1];
                     s[1] = image[i + 1][j - 1];
@@ -207,9 +207,9 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
             {
                 if (i == 0)
                 {
-                    BYTE red = 0;
-                    BYTE green = 0;
-                    BYTE blue = 0;
+                    double red = 0;
+                    double green = 0;
+                    double blue = 0;
 
                     s[0] = image[i][j - 1];
                     s[1] = image[i + 1][j - 1];
@@ -231,9 +231,9 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
                 }
                 else if (i == height - 1)
                 {
-                    BYTE red = 0;
-                    BYTE green = 0;
-                    BYTE blue = 0;
+                    double red = 0;
+                    double green = 0;
+                    double blue = 0;
 
                     s[0] = image[i][j - 1];
                     s[1] = image[i - 1][j - 1];
@@ -255,9 +255,9 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
                 }
                 else
                 {
-                    BYTE red = 0;
-                    BYTE green = 0;
-                    BYTE blue = 0;
+                    double red = 0;
+                    double green = 0;
+                    double blue = 0;
 
                     s[0] = image[i - 1][j - 1];
                     s[1] = image[i - 1][j];
